feat(textures): adopt core's TextureId, key the registry on it, add nativeTextureId
Phase 2 of labelle-gfx#328 (RFC-TEXTURE-ID-TYPING). Pins labelle-core
v1.28.0 for the types.

- `TextureId` is now core's type rather than a gfx-local duplicate, so
  gfx, the backends, the engine and games converge on ONE nominal type.
  `BackendTextureId` is re-exported alongside it.
- The registry is keyed on `TextureId` instead of a bare `u32`. This is
  what makes the guarantee structural: a BACKEND id is a different type
  and can no longer be used as a key by accident.
- New `nativeTextureId(TextureId) ?BackendTextureId` — the one legal
  conversion between the id spaces, on the engine and forwarded through
  `GfxRenderer`. Before it existed, callers handed a `TextureId` straight
  to a backend accessor and it compiled, which is how #326 silently
  blanked a downstream menu.

Core ships no `from(u32)` on purpose, so the ~11 `TextureId.from(...)`
call sites became explicit `@enumFromInt`. Two remain in src and both
are engine-facing seams (`drawMesh`, `drawScreenTexture`) that still
take a bare `u32`; typing those is phase 4. `mintTextureKey` keeps the
only allocation-site cast.

`nativeTextureId` currently retags `info.backend_texture.id` because
backends still declare `Texture.id: u32`. Marked SCAFFOLDING — phase 3
gives them `BackendTextureId` and the retag goes away.

Tests exercise the seam through BOTH the engine and `GfxRenderer` — a
forwarder that exists only on the inner type is a no-op in practice
(the gfx#291 lesson). Verified non-vacuous: breaking the wrapper
forwarder fails the suite. 147/147.

+++ src/retained_engine/bounds.zig
@@ -74,7 +74,7 @@ pub fn CullBounds(comptime Self: type) type {
             if (v.source_rect) |sr| {
                 display_w = if (sr.display_width > 0) sr.display_width else @abs(sr.width);
                 display_h = if (sr.display_height > 0) sr.display_height else @abs(sr.height);
-            } else if (self.textures.get(v.texture.toInt())) |t| {
+            } else if (self.textures.get(v.texture)) |t| {
                 display_w = t.width;
                 display_h = t.height;
             }
